// apps/web/src/shared/api/generated/model/loansControllerListFilter.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type LoansControllerListFilter =
  (typeof LoansControllerListFilter)[keyof typeof LoansControllerListFilter];

export const LoansControllerListFilter = {
  all: "all",
  return_soon: "return_soon",
  overdue: "overdue",
  no_return_date: "no_return_date",
  has_reminder: "has_reminder",
  without_reminder: "without_reminder",
} as const;